Moved helpers tests into test/helpers.
