+++ test/helpers/linkUnlinkFilter.js
@@ -1,7 +1,7 @@
 
-var linkFilter    = require('../lib/linkUnlinkFilter')
+var linkFilter    = require('../../lib/linkUnlinkFilter')
   , expect        = require("must")
-  , fixture       = require("./fixture")
+  , fixture       = require("../fixture")
 
 describe('filter couples of link/unlink and unlink/link', function() {
